Fix the silliest mistake
In my tabu search I sorted the wrong way round and thus considered the
allocations with lower payoff first. X( Correct that.

Also slightly tweak the formatting of the visualization.

Also add an input transformation that gives us sorted maps. It's no good
having the input in an unpredictable order when you're trying to figure
out why your algorithm is bad. Of course, since the algorithm doesn't
depend on names, it would make sense to have vectors as input, which
would also guarantee order. However, for now sorting the maps was
quicker to implement.